test: port close(code) event-sequence test to node:test so it runs on Node too
Use node:test + node:assert instead of bun:test so the same file runs
under both `bun test` and `node --test`, and add the self-check that
spawns node on the file (same pattern as node-http2-upgrade.test.mts).
All 8 cases pass on Node v26.3.0 and on this branch; 6 of 8 fail on main.
